chore: moving around scripts and adding support for the new testids
