+++ testing/cypress.config.ts
@@ -18,6 +18,7 @@ export default defineConfig({
   },
   e2e: {
     baseUrl: "https://bcgov.github.io/sso-requests-sandbox",
+    projectId: "gctfmh",
     setupNodeEvents(on, config) {
       // implement node event listeners here
     },
