+++ testing/cypress/e2e/dev-only/dev-deleterequest.cy.ts
@@ -1,6 +1,6 @@
 // This spec is used to delete created requests
 
-import Request from "../appActions/Request";
+import Request from "../../appActions/Request";
 let table = [];
 
 describe("Delete Request", () => {
