+++ testing/cypress/e2e/dev-only/dev-viewrequest.cy.ts
@@ -1,7 +1,7 @@
 // This spec is used to navigate to the request detail page for a specific request.
 // It is used to verify that the request detail page is not broken and to verify that basic navigation is possible.
 
-import Request from "../appActions/Request";
+import Request from "../../appActions/Request";
 let table = [];
 
 describe("View Request", () => {
