+++ testing/cypress/fixtures/createdRequest.json
@@ -1,14 +1,1 @@
-[
-  "00008948",
-  "00008953",
-  "00008954",
-  "00008955",
-  "00008956",
-  "00008957",
-  "00008958",
-  "00008974",
-  "00008975",
-  "00008976",
-  "00008977",
-  "00008978"
-]
+[]
