+++ testing/cypress/support/commands.ts
@@ -69,3 +69,4 @@ Cypress.Commands.add("assertValueCopiedToClipboard", (value) => {
     });
   });
 });
+
