+++ testing/cypress/support/global.d.ts
@@ -1,10 +1,10 @@
 declare namespace Cypress {
   interface Chainable<Subject> {
     login(
-        username: string, 
-        password: string, 
-        host: string, 
-        siteminder: string
+      username: string,
+      password: string,
+      host: string,
+      siteminder: string
     ): Chainable<any>;
 
     logout(host: string): void;
